requirements.txt
PyJWT
cryptography    
python-dotenv